requirements.txt
Django==4.2
python-dotenv==1.0.0
Requests==2.30.0
urllib3==2.0.2
